feat: support setting of python and poetry versions
This change introduces variables for allowing you to set whatever
versions of Python and poetry you'd like to install. You set these using
the GitHub Actions `with` block, where you can set `python_version` and
`poetry_version`. There are defaults listed in the action.yml. We are
considering letting those defaults automatically go lookup the latest
versions of python and/or poetry. This is as opposed to the previous
behavior which would only run the latest version of Python, and the
latest version of poetry.

At the same time, we've switched off using Alpine as this has been a
popular request.

Fixes #8
Fixes #9
Fixes #10