feat: Support under specified versions of Python
This change allows you to under specify a version of Python, such as
dropping the patch version number, and the action will still work. This
works even if you drop the minor version number and the patch version
number.

We're using the https://github.com/momo-lab/xxenv-latest plugin to find
the latest version of Python based on whatever version you gave it.

Fixes #12